perf(kernel): route GLM DSA scoring to Gluon
GLM 5.2 exposes scoring weights as a non-contiguous BF16 view. That shape did
not select the AMD Gluon scoring kernels and required the slower Triton path
with an FP32 weight conversion.

Accept BF16 scoring weights, compact them inside the Gluon implementation when
needed, and register BF16 weight signatures for decode and prefill selection.
Keep FP32 weight support and the existing public interface intact.

Verification on an AMD Instinct MI355X at upstream 8ae20e10:
- Changed-file pre-commit hooks passed. The all-files run reached isort, which
  rewrote four files introduced by upstream 8ae20e10 and untouched here; those
  unrelated edits were discarded rather than included in this branch.
- Decode compared 50,048 finite scores with maximum absolute error 4.47e-8,
  mean absolute error 5.13e-9, no mask violations, and exact sampled top-k
  overlap.
- Four prefill chunks had maximum absolute error at most 1.04e-7, no mask
  violations, and exact sampled top-k overlap.

Isolated GPU timing after two warmups:
- Decode over about 50K candidates took 71.901 us Gluon median over 20
  iterations, versus 174.582 us for Triton, a 2.43x speedup.
- Prefill with 8,192 queries in four memory-bounded chunks took 466.657 ms
  Gluon median total over 5 iterations, versus 647.813 ms for Triton, a 1.39x
  speedup. Both totals include weight preparation.

Model qualification used this commit's source tree
(a39c39176fa4de9faf0d15e7a0e89d872e852a09) before this metadata-only message
update. GLM 5.2 MXFP4 ran on four MI355X GPUs with attention and
mixture-of-experts tensor parallelism of four, FP8 key/value cache, standard
decoding, an 8,209-token input, and a two-token output cap. The single request
completed successfully in 51.215 seconds with 28,780.80 ms first-token latency
and 22,434.14 ms time per output token. Decode graph replay was observed on all
four ranks with graph padding disabled; prefill remained eager. This
single-request result is a qualification check, not a throughput benchmark.

Raw local evidence:
- claude_tmp/glm52/qualification/glm52-gluon-dsa-scoring__precommit__20260805T170700Z
- claude_tmp/glm52/qualification/glm52-gluon-dsa-scoring__kernel__20260805T171700Z
- claude_tmp/glm52/qualification/glm52-gluon-dsa-scoring__agentic-quick__20260805T172000Z

+++ tokenspeed-kernel/test/test_kernel_api_selection.py
@@ -1105,7 +1105,7 @@ def _attention_dsa_prefill_fp8_dense() -> object:
 
 def _attention_dsa_decode_topk() -> object:
     q = torch.empty((2, 2, 128), dtype=torch.bfloat16)
-    weights = torch.empty((2, 2), dtype=torch.float32)
+    weights = torch.empty((2, 2), dtype=torch.bfloat16)
     index_k = torch.zeros((128, 132), dtype=torch.uint8)
     seq_lens = torch.tensor([64, 64], dtype=torch.int32)
     block_table = torch.zeros((2, 1), dtype=torch.int32)
@@ -1128,7 +1128,7 @@ def _attention_dsa_prefill_topk(
     override: str | None = None,
 ) -> object:
     q = torch.empty((2, 2, 128), dtype=torch.bfloat16)
-    weights = torch.empty((2, 2), dtype=torch.float32)
+    weights = torch.empty((2, 2), dtype=torch.bfloat16)
     index_k = torch.zeros((128, 132), dtype=torch.uint8)
     kv_workspace_slots = torch.arange(64, dtype=torch.int64)
     row_starts = torch.tensor([0, 8], dtype=torch.int32)
